requirements.txt
kfp==2.7.0
google-cloud-pipeline-components==2.13.1
google-cloud-storage==2.16.0
openai==1.14.3
pydantic==2.6.4
langchain==0.1.13
langchain-openai==0.1.1
requests==2.28.1
python-dotenv==1.0.1